Refactor code folders.
