--- new_backend/deputies/parsers/__init__.py
@@ -1,1 +1,0 @@
-
